+++ src/main/java/ch/ethz/globis/phtree/v13/PhTree13.java
@@ -154,6 +154,9 @@ public class PhTree13<T> implements PhTree<T> {
 
 	@Override
 	public PhTreeStats getStats() {
+		if (getRoot() == null) {
+			return new PhTreeStats(DEPTH_64);
+		}
 		return getStats(0, getRoot(), new PhTreeStats(DEPTH_64));
 	}