+++ src/main/java/ch/ethz/globis/phtree/v16hd/PhTree16HD.java
@@ -174,6 +174,9 @@ public class PhTree16HD<T> implements PhTree<T> {
 
 	@Override
 	public PhTreeStats getStats() {
+		if (getRoot() == null) {
+			return new PhTreeStats(DEPTH_64);
+		}
 		return getStats(0, getRoot(), new PhTreeStats(DEPTH_64));
 	}